+++ src/lib.rs
@@ -19,7 +19,8 @@ pub mod scheduler;
 /// Clock synchronization utilities
 pub mod sync;
 
-pub use protocol::client::{ProtocolClient, ProtocolClientBuilder};
+pub use protocol::client::ProtocolClient;
+pub use protocol::client_builder::ProtocolClientBuilder;
 pub use protocol::messages::{ClientHello, ServerHello};
 pub use scheduler::AudioScheduler;
