+++ src/protocol/mod.rs
@@ -3,6 +3,8 @@
 
 /// WebSocket client implementation
 pub mod client;
+/// Builder for easy construction of the client
+pub mod client_builder;
 /// Protocol message type definitions and serialization
 pub mod messages;
